Fixed issues, also note that my original few commits are inside my project2 repo because I accidentally pushed them there

+++ AVLTreeDebug.cpp
@@ -56,7 +56,7 @@ int main() {
 	cout << endl;
 
 	// findRange
-	vector<int> rangeTest = tree.findRange("D", "W");
+	vector<size_t> rangeTest = tree.findRange("D", "W");
 	//70 68 82 75 77 86
 	for (auto val: rangeTest) {
 		cout << val << " ";
